User's register vaccination

+++ users/models.py
@@ -1,7 +1,6 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
 
-# Create your models here.
 
 class CustomUser(AbstractUser):
     first_name = models.CharField(max_length=30)
@@ -10,7 +9,7 @@ class CustomUser(AbstractUser):
     emergency_contact = models.CharField(max_length=30, blank=True, null=True)
     gender = models.CharField(max_length=6)
     birthdate = models.DateField(blank=True, null=True)
-    age = models.IntegerField(default=0)
+    age = models.FloatField(default=0.0)
 
     def __str__(self):
         return self.email
